backend/pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = "Fastapi microblog backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "pydantic>=2.8.0,<3.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "PyJWT>=2.9.0,<3.0.0",
    "python-jose>=3.3.0,<4.0.0",
    "sqlmodel>=0.0.27,<1.0.0",
    "SQLAlchemy>=2.0.45,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "ruff>=0.6.9",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "faker>=28.0.0",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true

[tool.mypy]
python_version = "3.13"
disable_error_code = ["no-any-return", "return-value", "var-annotated"]
warn_unused_configs = true
disallow_untyped_calls = false
disallow_untyped_defs = false
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
ignore_missing_imports = true
strict_equality = true
plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing --cov-report=xml -p no:warnings"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]



